clean up stuff for merge and move ThTick_Drive to WIP
