+++ decompile/General/BOTS/BOTS_Defrag.c
@@ -1,5 +1,7 @@
 #include <common.h>
 
+//don't use this until all functions are decompiled and all calls are made to DECOMP_ variants.
+
 #include "BOTS_00_SetGlobalNavData.c"
 #include "BOTS_01_InitNavPath.c"
 #include "BOTS_02_EmptyFunc.c"
