+++ include/decomp_functions.h
@@ -898,3 +898,4 @@ void DECOMP_GAMEPAD_JogCon2(struct Driver* d, char val, short timeMS);
 void DECOMP_CAM_EndOfRace(struct CameraDC* cDC, struct Driver* d);
 u_int DECOMP_BOTS_ChangeState(struct Driver* driverVictim, int damageType, struct Driver* driverAttacker, int reason);
 void DECOMP_BOTS_Killplane(struct Thread* botThread);
+void DECOMP_COLL_FIXED_BotsSearch(short* posCurr, short* posPrev, struct ScratchpadStruct* sps);
